Adding a new class to fmidb and changing class inheritance hierarchy.
With this change all database classes inherit from a single parent, NFmiDatabase.
This is beneficial when combined with the new database class.

A new class NFmiPostgreSQL is added to fmidb. This class uses Postgres native
c++ -library to access the database. I was forced to do this since radon deadlocking
issues (STU-2066) are large related (in my opinion) to the ODBC driver which does some 
transaction fiddling behind the scenes. So the new class will work better, but of course 
it will addition code to maintain.

Combining the two changes we can (later) change NFmiRadonDB to not use inheritance but
composition: it will have one class member NFmiDatabase that will be either postgres
database or sqlite database (backend support still missing). This also allows to write
database-type specific sql queries which will be an improvement (ODBC driver does not tell
the backend type!)

Old classes NFmiOracle and NFmiODBC are changed to inherit from NFmiDatabase, but
otherwise they will work as before and can be used also in future projects. The real 
benefit of using NFmiPostgreSQL will only materialize when it's used for writing in several 
parallel threads.