Add initial support for fetching geometry information from projection-specific tables. So instead of one "geom" table, we will later have a different table for each geometry/projection. This will improve the current situation since currently all three used projections (latlon, rotlatlon, polster) are stored in one table that does not recognize the distinct features of each projection. Also when a fourth one (reduced gaussian grid) is added, this table does not fit any more. The downside of this change is that the table structure will become slightly more complex.
So this commit will add support for reduced gaussian grid
geometries which are stored in a separate table which exists
only in development database at the moment. The fetching
of other projections has not been changed for now, since the 
corresponding tables have not been created yet. So everything
should work as it did in the past.

Also file has been formatted with clang-formatter:

clang-format-3.8 -style="{BasedOnStyle: google, \
  BreakBeforeBraces: Allman, ColumnLimit: 120, \ 
  IndentWidth: 4, IndentCaseLabels: true, \
  UseTab: ForIndentation, TabWidth: 4}" \
  -dump-config