Add cache warming for grib1 parameter information.
When large grib files (hundreds or thousands of messages) are loaded,
grid_to_neons makes a lot of database queries. These is because for
each parameter/level combination we have to check what is the parameter
name.

Adding now a function that _can_ be called before starting to load data
that will "warm up the cache", meanining it will populate as much of the
cache as possible so that query count will be smaller. The warming is
"adaptive" as it doesn't only fetch all metadata from database, but also
adds cache entries for distinct hybrid levels even if no specific hybrid
level setting is found from database.

When no parameter information is found for a hybrid level, we also fill
the cache with empty data for all hybrid levels for this certain parameter.

Conflicts:
	source/NFmiRadonDB.cpp